iPad Version 🎉📱
Today is a new day, is a new day for the game of go. We work hard, we
drank a lot of coffee and took a lot of other things to offer you
ladies and gentlemen : the very first version of the Go game on iPad !
AMAZING.

On the program:
- Wonderful Splash screen 👙
- Amazing App Icon
- A square grid with full online too pretty. Waouh.
- A carousel that works with the finger. Yes, yes, on iPad. How is this
possible? We have made technological feat for you.
- Pawns of incredible colors. Orange & White.

Although there are still some bugs, we have coffee, cigarettes and the
day is still so up, patience, we coming soon ! Bisou, bisou XoXo 💗

The GoGame Team